University_Chatbot/requirements.txt

deepeval==3.5.9
dotenv==0.9.9
faiss-cpu==1.11.0
fastapi==0.115.13
google-generativeai
gradio==5.34.2
langchain-community==0.3.24
langchain-core==0.3.76
langchain-google-genai
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langgraph
langsmith==0.3.45
matplotlib==3.10.3
numpy==2.2.6
pandas==2.3.0
PyMuPDF==1.26.0
python-dotenv==1.1.1
requests==2.32.3
regex==2024.11.6
sentence-transformers==4.1.0

